UI Redesign and Folder Structure
Finished UI redesign on Tasks.
Tasks are now in order of latest.
Re-added different categories on sidepanel.
