--- components/logo.tsx
@@ -1,5 +1,0 @@
-export default function Logo() {
-  return (
-    <h1 className="font-extrabold text-4xl pt-6 text-center">Task Forge</h1>
-  );
-}
