+++ components/side-panel/side-panel-navigation.tsx
@@ -10,16 +10,16 @@ import {
   Plus,
   Square,
 } from "lucide-react";
-import SidePanelButton from "./side-panel-button";
 import { useRouter, useSearchParams } from "next/navigation";
+import { useState } from "react";
+import { categories } from "../providers/session-provider";
+import CategoryNew from "../task/category-new";
 import {
   Collapsible,
   CollapsibleContent,
   CollapsibleTrigger,
 } from "../ui/collapsible";
-import { useState } from "react";
-import CategoryNew from "../category-new";
-import { categories } from "../session-provider";
+import SidePanelButton from "./side-panel-button";
 
 export default function SidePanelNavigation() {
   const router = useRouter();
@@ -30,67 +30,6 @@ export default function SidePanelNavigation() {
   const [showNewCat, setShowNewCat] = useState(false);
   return (
     <div className="flex flex-col w-full pt-6">
-      <SidePanelButton
-        current={
-          tasks === "all" ||
-          (tasks === null && searchParams.get("settings") === null)
-        }
-        title={"Tasks"}
-        icon={
-          <Home
-            color={
-              tasks === "all" ||
-              (tasks === null && searchParams.get("settings") === null)
-                ? "#ffffff"
-                : "#000000"
-            }
-          />
-        }
-        onClick={() => router.push("/?tasks=all")}
-      />
-      <Collapsible
-        open={isOpen}
-        onOpenChange={setIsOpen}
-        className="w-full z-0"
-      >
-        <CollapsibleTrigger className="w-full">
-          <SidePanelButton
-            current={tasks === "custom-categories"}
-            title={"Categories"}
-            icon={isOpen ? <FolderOpen /> : <Folder />}
-          />
-        </CollapsibleTrigger>
-        <CollapsibleContent className="max-h-[32rem] overflow-y-auto overflow-x-hidden">
-          {true && (
-            <>
-              {categories.map((cat: Category) => (
-                <SidePanelButton
-                  key={cat.id}
-                  current={category === cat.id.toString()}
-                  title={cat.title}
-                  icon={<Folder color={cat.colour} />}
-                  onClick={() =>
-                    router.push("/?tasks=all-cat&category=" + cat.id)
-                  }
-                  category={cat}
-                />
-              ))}
-            </>
-          )}
-          <SidePanelButton
-            current={false}
-            title={"Add New Category"}
-            icon={<Plus />}
-            className="pl-6 hover:bg-muted text-sm"
-            onClick={() => setShowNewCat(true)}
-          />
-        </CollapsibleContent>
-      </Collapsible>
-      {showNewCat && <CategoryNew show={setShowNewCat} newCategory />}
-    </div>
-  );
-  return (
-    <div className="flex flex-col w-full">
       <SidePanelButton
         current={
           tasks === "all" ||
@@ -148,7 +87,6 @@ export default function SidePanelNavigation() {
                   onClick={() =>
                     router.push("/?tasks=all-cat&category=" + cat.id)
                   }
-                  className={"translate-x-5"}
                   category={cat}
                 />
               ))}
@@ -156,9 +94,9 @@ export default function SidePanelNavigation() {
           )}
           <SidePanelButton
             current={false}
-            title={"New Category"}
+            title={"Add New Category"}
             icon={<Plus />}
-            className={"translate-x-5"}
+            className="pl-6 hover:bg-muted text-sm"
             onClick={() => setShowNewCat(true)}
           />
         </CollapsibleContent>
